YD-166 Moved to ActivityService, fixed test
